Functions

Function Reference

Function Reference

Caesar

Index

Caesar

Caesar.loadmodel — Function.

mdl = loadmodel(model, [color=, offset=])

Brings model data into context, but generating or loading meshes, Geometries, etc as required by requested model, and must still be drawn with the Visualizer object using mdl(vis).

source

Caesar.getPoseExVertexNeoIDs — Function.

getPoseExVertexNeoIDs(neo4j.connection)

Return array of tuples with ExVertex IDs and Neo4j IDs for all poses.

source

Caesar.askneo4jcredentials! — Function.

askneo4jcredentials!(;addrdict::Dict{AbstractString, AbstractString})

Obtain Neo4j global database address and login credientials from STDIN, then insert and return in the addrdict colletion.

source

RoME.getRangeKDEMax2D — Function.

getRangeKDEMax2D(cgl::CloudGraph, session::AbstractString, vsym1::Symbol, vsym2::Symbol)

Calculate the cartesian distange between two vertices in the graph, by session and symbol names, and by maximum belief point.

source

Caesar.getLandmOtherSessNeoIDs — Function.

getLandmOtherSessNeoIDs{T <: AbstractString}(::CloudGraph, session::T="", multisessions=Vector{T}())

Return dict of dict of Neo4j vertex IDs by session and landmark symbols.

source

Caesar.db2jld — Function.

db2jld(cgl::CloudGraph, session::AbstractString, filename::AbstractString)

Fetch and save a FactorGraph session to a jld, using CloudGraph object and session definition.

source

db2jld(filename::AbstractString; addrdict::VoidUnion{Dict{AbstractString, AbstractString}}=nothing )

Fetch and save a FactorGraph session to a jld, using or asking STDIN for credentials in the addrdict field.

source

Caesar.identitypose6fg — Function.

identitypose6fg()

Initialize a and exiting or new factor graph with a first pose and prior as specified by default keywords or user.

source

Caesar.fetchrobotdatafirstpose — Function.

fetchrobotdatafirstpose(cg::CloudGraph, session::AbstractString)

Return dict of JSON parsed "robot_description" field as was inserted by counterpart insertrobotdatafirstpose! function. Used for storing general robot specific data in easily accessible manner.

source

Caesar.executeQuery — Function.

executeQuery(cg::CloudGraph, query::AbstractString)

Run Neo4j Cypher queries on the cloudGraph database, andreturn Tuple with the unparsed (results, loadresponse).

source

IncrementalInference.ls — Function.

ls(cgl::CloudGraph, session::AbstractString; sym::Symbol=, neoid::Int=,exvid::Int=)

List neighbors to node in cgl::CloudGraph by returning Dict{Sym}=(exvid, neoid, Symbol[labels]), and can take any of the three as input node identifier. Not specifying an identifier will result in all Variable nodes being returned.

source

Caesar.updateparallelposes! — Function.

updateparallelposes!(vis, poseswithdepth, wTb=::CoordinateTransformations.AbstractAffineMap)

Update all triads listed in poseswithdepth[Symbol(vert.label)] with wTb. Prevents cycles in remote tree viewer of DrakeVisualizer.

source

Caesar.removeReinsertMultisessionPrior! — Function.

removeReinsertMultisessionPrior!{T <: FunctorSingleton}(fgl::FactorGraph, exims::Dict{Symbol, Int}, prp2::T, sym::Symbol, uid::Int)

Return new multisession factor, symbol sym, inserted using prp2::PriorPoint2DensityNH using user ExVertex id (uid), after checking existing multisession dict (exims::(exvsym=>neoid)) if any nodes should be removed before inserting new ones on that graph node. Assuming just one multisession prior per node.

source

Caesar.appendvertbigdata! — Function.

appendvertbigdata!(cloudGraph, cloudvert, descr, data)

Append big data element into current blob store and update associated global vertex information.

source

appendvertbigdata!(fg, vert, descr, data)

Append big data element into current blob store and update associated global vertex information.

source

appendvertbigdata!(fg, sym, descr, data)

Append big data element into current blob store using parent appendvertbigdata!, but here specified by symbol of variable node in the FactorGraph. Note the default data layer api definition. User must define dlapi to refetching the vertex from the data layer. localapi avoids repeated network database fetches.

source

Caesar.parseMergeVertAttr! — Function.

parseMergeVertAttr(v, elem)

Parse elem dictionary according to thin frtend interface from other languages, and merge contents into attributes of v::Graphs.ExVertex.

source

Caesar.updatenewverts! — Function.

updatenewverts!(fgl::FactorGraph; N::Int)

Convert vertices of session in Neo4j DB with Caesar.jl's required data elements in preparation for MM-iSAMCloudSolve process.

source

Caesar.mergeCloudVertex! — Function.

mergeCloudVertex!(...)

Hodgepodge function to merge data in CloudVertex

source

Caesar.getVertNeoIDs! — Function.

getVertNeoIDs!(::CloudGraph, res::Dict{Symbol, Int}; session::AbstractString="NA")

Insert into and return dict res with Neo4j IDs of ExVertex labels as stored per session in Neo4j database.

source

Caesar.whosNear3D — Function.

whosNear3D(cg::CloudGraph, session::AbstractString; x,y,z,roll, pitch,yaw,dist,angle )

Find vertices near the point specified and return dictionary of symbol to Neo4j ID pairs.

source

Caesar.getprpt2kde — Function.

getprp2kde(::CloudGraph, neoids::Vector{Int}; N::Int=100)

Return PriorPoint2DensityNH with N points based on beliefs of neoids, and equal share null hypothesis between length(neoids)+1 beliefs.

source

Caesar.whosNear2D — Function.

whosNear2D(cg::CloudGraph, session::AbstractString; x,y,yaw,dist,angle )

Find vertices near the point specified and return dictionary of symbol to Neo4j ID pairs.

source

Caesar.hasBigDataElement — Function.

hasBigDataElement(vertex, description)

Return true if vertex has bigDataElements with matching description.

source

Caesar.getAllLandmarkNeoIDs — Function.

getAllLandmarkNeoIDs(::Dict{Symbol, Dict{Symbol, Int}}, ::Symbol)

Return Vector{Int} of Neo4j vertex IDs relating to symbol, as listed in lm2others.

source

Caesar.consoleaskuserfordb — Function.

consoleaskuserfordb(;nparticles=false, drawdepth=false, clearslamindb=false)

Obtain database addresses and login credientials from STDIN, as well as a few case dependent options.

source

Caesar.drawAllOdometryEdges! — Function.

drawAllOdometryEdges!(fgl::FactorGraph, fr::Symbol, to::Symbol; scale, color, api  )

Assume odometry chain and draw edges between subsequent poses. Use keyword arguments to change colors, etc.

source

Caesar.standardcloudgraphsetup — Function.

standardcloudgraphsetup(;addrdict=nothing, nparticles=false, drawdepth=false, clearslamindb=false)

Connect to databases via network according to addrdict, or ask user for credentials and return active cloudGraph object, as well as addrdict.

source

Caesar.resetentireremotesession — Function.

resetentireremotesession(conn, session)

match (n:session) remove n.backendset, n.ready, n.data, n.bigData, n.label, n.packedType, n.exVertexId, n.shape, n.width set n :NEWDATA return n

source

Caesar.getfirstpose — Function.

getfirstpose(cg::CloudGraph, session::AbstractString)

Return Tuple{Symbol, Int} of first pose symbol and Neo4j node ID.

source

Caesar.getExVertexNeoIDs — Function.

getExVertexNeoIDs(neo4j.connection, label="", session="")

Return array of tuples with ExVertex IDs and Neo4j IDs for vertices with label in session.

source

Caesar.findExistingMSConstraints — Function.

findExistingMSConstraints(fgl::FactorGraph)

Return Dict{Symbol, Int} of vertex symbol to Neo4j node ID of MULTISESSION constraints in this fgl.sessionname.

source

Caesar.removeNeo4jID — Function.

removeNeo4jID(cg::CloudGraph, neoid=-1)

Remove node from Neo4j according to Neo4j Node ID. Big data elements that may be associated with this node are not removed.

source

Caesar.drawLineBetween! — Function.

drawLineBetweenPose3(fr::Graphs.ExVertex, to::Graphs.ExVertex; scale=, color=  )

Draw a line segment between to vertices.

source

drawLineBetween(fgl::FactorGraph, fr::Symbol, to::Symbol; scale, color, api  )

Draw a line segment between to nodes in the factor graph.

source

Caesar.getBigDataElement — Function.

getBigDataElement(vertex::CloudVertex, description)

Walk through vertex bigDataElements and return the last matching description.

source

Caesar.askmongocredentials! — Function.

askmongocredentials!(addrdict=Dict{AbstractString, AbstractString})

Obtain Mongo database address and login credientials from STDIN, then insert and return in the addrdict colletion.

source

Caesar.rmInstMultisessionPriors! — Function.

rmInstMultisessionPriors!{T <: AbstractString}(::CloudGraph; session::T=, multisessions:Vector{T}= )

source

RoME.getLastPose — Function.

getLastPose(cg::CloudGraph, session::AbstractString)

Return Tuple{Symbol, Int} of first pose symbol and Neo4j node ID.

source

Caesar.insertrobotdatafirstpose! — Function.

insertrobotdatafirstpose!(cg::CloudGraph, session::AbstractString, robotdict::Dict)

Saves robotdict via JSON to first pose in a SESSION in the database. Used for storing general robot specific data in easily accessible manner. Can fetch later retrieve same dict with counterpart fetchrobotdatafirstpose function.

source

Caesar.fetchsubgraph! — Function.

fetchsubgraph!(::FactorGraph, ::Vector{CloudVertex}, numneighbors::Int=0)

Fetch and insert list of CloudVertices into FactorGraph object, up to neighbor depth.

source

fetchsubgraph!(::FactorGraph, ::Vector{Int}, numneighbors::Int=0)

Fetch and insert list of Neo4j IDs into FactorGraph object, up to neighbor depth.

source

Caesar.getLocalSubGraphMultisession — Function.

getLocalSubGraphMultisession{T <: AbstractString}(cg::CloudGraph, lm2others; session::T="", numneighbors::Int=0)

Return subgraph copy of type FactorGraph contaning values from session in lm2others, and Vector{Symbol} of primary key symbols used for graph exstraction.

source

Caesar.getnewvertdict — Function.

getnewvertdict(conn, session)

Return a dictionary with frtend and mongo_keys json string information for :NEWDATA elements in Neo4j database.

source

Index

Caesar.appendvertbigdata!

Caesar.askmongocredentials!

Caesar.askneo4jcredentials!

Caesar.consoleaskuserfordb

Caesar.db2jld

Caesar.drawAllOdometryEdges!

Caesar.drawLineBetween!

Caesar.executeQuery

Caesar.fetchrobotdatafirstpose

Caesar.fetchsubgraph!

Caesar.findExistingMSConstraints

Caesar.getAllLandmarkNeoIDs

Caesar.getBigDataElement

Caesar.getExVertexNeoIDs

Caesar.getLandmOtherSessNeoIDs

Caesar.getLocalSubGraphMultisession

Caesar.getPoseExVertexNeoIDs

Caesar.getVertNeoIDs!

Caesar.getfirstpose

Caesar.getnewvertdict

Caesar.getprpt2kde

Caesar.hasBigDataElement

Caesar.identitypose6fg

Caesar.insertrobotdatafirstpose!

Caesar.loadmodel

Caesar.mergeCloudVertex!

Caesar.parseMergeVertAttr!

Caesar.removeNeo4jID

Caesar.removeReinsertMultisessionPrior!

Caesar.resetentireremotesession

Caesar.rmInstMultisessionPriors!

Caesar.standardcloudgraphsetup

Caesar.updatenewverts!

Caesar.updateparallelposes!

Caesar.whosNear2D

Caesar.whosNear3D

IncrementalInference.ls

RoME.getLastPose

RoME.getRangeKDEMax2D