Visualization Function Reference

Visualization Function Reference

RoMEPlotting

RoMEPlotting

Missing docstring.

Missing docstring for drawLandms. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drawPoses. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drawPosesLandms. Check Documenter's build log for details.

KernelDensityEstimatePlotting.plotKDE — Function

plotKDE(fgl, sym; solveKey, dims, title, levels, fill, layers, c, overlay)

A peneric KDE plotting function that allows marginals of higher dimensional beliefs and various keyword options.

Example:

p = kde!(randn(3,100))

plotKDE(p)
plotKDE(p, dims=[1;2], levels=3)
plotKDE(p, dims=[1])

q = kde!(5*randn(3,100))
plotKDE([p;q])
plotKDE([p;q], dims=[1;2], levels=3)
plotKDE([p;q], dims=[1])

RoMEPlotting.plotPose — Function

plotPose(pt, pp)
plotPose(pt, pp, title; levels, c, axis, scale, overlay, hdl)

Plot pose belief as contour information on visually sensible manifolds.

plotPose(fgl, syms; solveKey, levels, c, axis, scale, show, filepath, app, hdl)

Example: pl = plotPose(fg, [:x1; :x2; :x3])

RoMEPlotting.plotProductVsKDE — Function

Analysis function to compare KDE plots between the factor graph centric product of a variable with current value stored in the factor graph object.

WORK IN PROGRESS Not all functions have been added to this directory yet.