Available Variables/Factors

Variables and Factors Already Available in Caesar

Variables Available in Caesar

You can check for the latest variable types by running the following in your terminal:

using RoME, Caesar
subtypes(IncrementalInference.InferenceVariable)

Note: This has been made available as IncrementalInference.getCurrentWorkspaceVariables() in IncrementalInference v0.4.4.

The current list of available variable types is:

RoME.Point2 - A 2D coordinate consisting of [x, y, theta]

RoME.Pose2 - A 2D coordinate and a rotation (i.e. bearing) consisting of [x, y, z, and theta]

RoME.DynPoint2 - A 2D coordinate and linear velocities

RoME.DynPose2 - A 2D coordinate, linear velocities, and a rotation

RoME.Point3 - A 3D coordinate consisting of [x, y, z]

RoME.Pose3 - A 3D coordinate and 3 associated rotations consisting of [x, y, z, theta, phi, psi]

RoME.InertialPose3 - A 3D coordinate and rotation pose along with velocity and IMU bias calibration terms

Note several more variable and factors types have been implemented which will over time be incorporated into standard RoME release. Please open an issue with JuliaRobotics/RoME.jl for specific requests, problems, or suggestions. Contributions are also welcome.

Factors Available in Caesar

You can check for the latest factor types by running the following in your terminal:

using RoME, Caesar
println("- Singletons (priors): ")
println.(sort(string.(subtypes(IncrementalInference.FunctorSingleton))));
println("- Pairwise (variable constraints): ")
println.(sort(string.(subtypes(IncrementalInference.FunctorPairwise))));
println("- Pairwise (variable minimization constraints): ")
println.(sort(string.(subtypes(IncrementalInference.FunctorPairwiseMinimize))));

Existing prior (unary) factors in Caesar.jl/RoME.jl/IIF.jl include:

IncrementalInference.Prior — Type.

struct Prior{T} <: FunctorSingleton

RoME.PriorPoint2 — Type.

mutable struct PriorPoint2{T} <: FunctorSingleton

Direction observation information of a Point2 variable.

RoME.PriorPose2 — Type.

mutable struct PriorPose2{T} <: FunctorSingleton

Introduce direct observations on all dimensions of a Pose2 variable:

Example:

PriorPose2( MvNormal([10; 10; pi/6.0], Matrix(Diagonal([0.1;0.1;0.05].^2))) )

RoME.PriorPolar — Type.

mutable struct PriorPolar{T1<:Union{AliasingScalarSampler, BallTreeDensity, Distribution}, T2<:Union{AliasingScalarSampler, BallTreeDensity, Distribution}} <: FunctorSingleton

RoME.PriorPoint3 — Type.

mutable struct PriorPoint3{T} <: FunctorSingleton

Direction observation information of a Point3 variable.

RoME.PriorPose3 — Type.

mutable struct PriorPose3 <: FunctorSingleton

Direct observation information of Pose3 variable type.

Existing n-ary factors in Caesar.jl/RoME.jl/IIF.jl include:

RoME.Point2Point2 — Type.

mutable struct Point2Point2{D<:Union{AliasingScalarSampler, BallTreeDensity, Distribution}} <: FunctorPairwise

RoME.Point2Point2WorldBearing — Type.

mutable struct Point2Point2WorldBearing{T} <: FunctorPairwise

RoME.Pose2Point2Bearing — Type.

mutable struct Pose2Point2Bearing{B<:Union{AliasingScalarSampler, BallTreeDensity, Distribution}} <: FunctorPairwiseMinimize

Single dimension bearing constraint from Pose2 to Point2 variable.

RoME.Pose2Point2BearingRange — Type.

mutable struct Pose2Point2BearingRange{B<:Union{AliasingScalarSampler, BallTreeDensity, Distribution}, R<:Union{AliasingScalarSampler, BallTreeDensity, Distribution}} <: FunctorPairwise

Bearing and Range constraint from a Pose2 to Point2 variable.

RoME.Pose2Point2Range — Type.

mutable struct Pose2Point2Range{T} <: FunctorPairwise

Range only measurement from Pose2 to Point2 variable.

RoME.Pose2Pose2 — Type.

mutable struct Pose2Pose2{T} <: FunctorPairwise

RoME.Pose3Pose3 — Type.

mutable struct Pose3Pose3 <: FunctorPairwise

RoME.InertialPose3 — Type.

mutable struct InertialPose3 <: FunctorPairwise

Inertial Odometry version of preintegration procedure and used as a factor between InertialPose3 types for inertial navigation in factor graphs.

Extending Caesar with New Variables and Factors

A question that frequently arises is how to design custom variables and factors to solve a specific type of graph. One strength of Caesar is the ability to incorporate new variables and factors at will. Please refer to Adding Factors for more information on creating your own factors.