Function Reference

Function Reference

Function Reference

Caesar

Caesar

Caesar.getPoseExVertexNeoIDs — Function.

getPoseExVertexNeoIDs(conn; ready, backendset, session, reqbackendset)

Return array of tuples with ExVertex IDs and Neo4j IDs for all poses.

source

Caesar.getLandmOtherSessNeoIDs — Function.

getLandmOtherSessNeoIDs{T <: AbstractString}(::CloudGraph, session::T="", robot::T="", user::T="", multisessions=Vector{T}())

Return dict of dict of Neo4j vertex IDs by session and landmark symbols.

source

Caesar.db2jld — Function.

db2jld(cgl::CloudGraph, session::AbstractString, filename::AbstractString)

Fetch and save a FactorGraph session to a jld, using CloudGraph object and session definition.

source

db2jld(filename::AbstractString; addrdict::NothingUnion{Dict{AbstractString, AbstractString}}=nothing )

Fetch and save a FactorGraph session to a jld, using or asking STDIN for credentials in the addrdict field.

source

Caesar.fetchrobotdatafirstpose — Function.

fetchrobotdatafirstpose(cg::CloudGraph, session::AbstractString, robot::AbstractString, user::AbstractString)

Return dict of JSON parsed "robot_description" field as was inserted by counterpart insertrobotdatafirstpose! function. Used for storing general robot specific data in easily accessible manner.

source

Caesar.executeQuery — Function.

executeQuery(connection, query)

Run Neo4j Cypher queries on the cloudGraph database, and return Tuple with the unparsed (results, loadresponse).

source

Caesar.appendvertbigdata! — Function.

appendvertbigdata!(cloudGraph, cv, description, data)

Append big data element into current blob store and update associated global vertex information.

source

appendvertbigdata!(fgl, vert, description, data)

Append big data element into current blob store and update associated global vertex information.

source

appendvertbigdata!(fg, sym, descr, data)

Append big data element into current blob store using parent appendvertbigdata!, but here specified by symbol of variable node in the FactorGraph. Note the default data layer api definition. User must define dlapi to refetching the vertex from the data layer. localapi avoids repeated network database fetches.

source

Caesar.updatenewverts! — Function.

updatenewverts!(fgl::FactorGraph; N::Int)

Convert vertices of session in Neo4j DB with Caesar.jl's required data elements in preparation for MM-iSAMCloudSolve process.

source

Caesar.getVertNeoIDs! — Function.

getVertNeoIDs!(::CloudGraph, res::Dict{Symbol, Int}; session::AbstractString="NA", robot::AbstractString="NA", user::AbstractString="NA")

Insert into and return dict res with Neo4j IDs of ExVertex labels as stored per session in Neo4j database.

source

Caesar.whosNear3D — Function.

whosNear3D(cg, session, robot, user; x, y, z, roll, pitch, yaw, dist, angle)

Find vertices near the point specified and return dictionary of symbol to Neo4j ID pairs.

source

Caesar.getprpt2kde — Function.

getprp2kde(::CloudGraph, neoids::Vector{Int}; N::Int=100)

Return PriorPoint2DensityNH with N points based on beliefs of neoids, and equal share null hypothesis between length(neoids)+1 beliefs.

source

Caesar.whosNear2D — Function.

whosNear2D(cg, session, robot, user; x, y, yaw, dist, angle)

Find vertices near the point specified and return dictionary of symbol to Neo4j ID pairs.

source

Caesar.hasBigDataElement — Function.

hasBigDataElement(vertex, description)

Return true if vertex has bigDataElements with matching description.

source

Caesar.getAllLandmarkNeoIDs — Function.

getAllLandmarkNeoIDs(::Dict{Symbol, Dict{Symbol, Int}}, ::Symbol)

Return Vector{Int} of Neo4j vertex IDs relating to symbol, as listed in lm2others.

source

Caesar.consoleaskuserfordb — Function.

consoleaskuserfordb(; nparticles, drawdepth, clearslamindb, multisession, drawedges)

Obtain database addresses and login credientials from STDIN, as well as a few case dependent options.

source

Caesar.standardcloudgraphsetup — Function.

standardcloudgraphsetup(; addrdict, nparticles, drawdepth, drawedges, clearslamindb, multisession)

Connect to databases via network according to addrdict, or ask user for credentials and return active cloudGraph object, as well as addrdict.

source

Caesar.resetentireremotesession — Function.

resetentireremotesession(conn, session, robot, user)

match (n:session) remove n.backendset, n.ready, n.data, n.bigData, n.label, n.packedType, n.exVertexId, n.shape, n.width set n :NEWDATA return n

source

Caesar.getfirstpose — Function.

getfirstpose(cg::CloudGraph, session::AbstractString, robot::AbstractString, user::AbstractString)

Return Tuple{Symbol, Int} of first pose symbol and Neo4j node ID.

source

Caesar.getExVertexNeoIDs — Function.

getExVertexNeoIDs(conn; label, ready, backendset, session, robot, user, reqbackendset)

Return array of tuples with ExVertex IDs and Neo4j IDs for vertices with label in session.

source

Caesar.findExistingMSConstraints — Function.

findExistingMSConstraints(fgl::FactorGraph)

Return Dict{Symbol, Int} of vertex symbol to Neo4j node ID of MULTISESSION constraints in this fgl.sessionname.

source

Caesar.removeNeo4jID — Function.

removeNeo4jID(cg::CloudGraph, neoid=-1)

Remove node from Neo4j according to Neo4j Node ID. Big data elements that may be associated with this node are not removed.

source

Caesar.getBigDataElement — Function.

getBigDataElement(vertex, description)

Walk through vertex bigDataElements and return the last matching description.

source

Caesar.rmInstMultisessionPriors! — Function.

rmInstMultisessionPriors!(::CloudGraph; session<:AbstractString=, multisessions::Vector{<:AbstractString}= )

source

Caesar.insertrobotdatafirstpose! — Function.

insertrobotdatafirstpose!(cg::CloudGraph, session::AbstractString, robot::AbstractString, user::AbstractString, robotdict::Dict)

Saves robotdict via JSON to first pose in a SESSION in the database. Used for storing general robot specific data in easily accessible manner. Can fetch later retrieve same dict with counterpart fetchrobotdatafirstpose function.

source

Caesar.fetchsubgraph! — Function.

fetchsubgraph!(::FactorGraph, ::Vector{CloudVertex}, numneighbors::Int=0)

Fetch and insert list of CloudVertices into FactorGraph object, up to neighbor depth.

source

fetchsubgraph!(::FactorGraph, ::Vector{Int}, numneighbors::Int=0)

Fetch and insert list of Neo4j IDs into FactorGraph object, up to neighbor depth.

source

Caesar.getLocalSubGraphMultisession — Function.

getLocalSubGraphMultisession{T <: AbstractString}(cg::CloudGraph, lm2others; session::T="", numneighbors::Int=0)

Return subgraph copy of type FactorGraph contaning values from session in lm2others, and Vector{Symbol} of primary key symbols used for graph exstraction.

source

Caesar.getnewvertdict — Function.

getnewvertdict(conn, session::AbstractString, robot::AbstractString, user::AbstractString)

Return a dictionary with frtend and mongo_keys json string information for :NEWDATA elements in Neo4j database.

source