Caesar's Reference

Function Reference

Function Reference

Caesar

RoME

IncrementalInference

WORK IN PROGRESS Not all functions have been added to this directory yet.

Caesar

Caesar.getPoseExVertexNeoIDs — Function.

getPoseExVertexNeoIDs(conn; ready, backendset, session, reqbackendset)

Return array of tuples with ExVertex IDs and Neo4j IDs for all poses.

source

Caesar.getLandmOtherSessNeoIDs — Function.

getLandmOtherSessNeoIDs{T <: AbstractString}(::CloudGraph, session::T="", robot::T="", user::T="", multisessions=Vector{T}())

Return dict of dict of Neo4j vertex IDs by session and landmark symbols.

source

Caesar.db2jld — Function.

db2jld(cgl::CloudGraph, session::AbstractString, filename::AbstractString)

Fetch and save a FactorGraph session to a jld, using CloudGraph object and session definition.

source

db2jld(filename::AbstractString; addrdict::NothingUnion{Dict{AbstractString, AbstractString}}=nothing )

Fetch and save a FactorGraph session to a jld, using or asking STDIN for credentials in the addrdict field.

source

Caesar.fetchrobotdatafirstpose — Function.

fetchrobotdatafirstpose(cg::CloudGraph, session::AbstractString, robot::AbstractString, user::AbstractString)

Return dict of JSON parsed "robot_description" field as was inserted by counterpart insertrobotdatafirstpose! function. Used for storing general robot specific data in easily accessible manner.

source

Caesar.executeQuery — Function.

executeQuery(connection, query)

Run Neo4j Cypher queries on the cloudGraph database, and return Tuple with the unparsed (results, loadresponse). Throws an error if the query fails.

source

Caesar.appendvertbigdata! — Function.

appendvertbigdata!(cloudGraph, cv, description, data)

Append big data element into current blob store and update associated global vertex information.

source

appendvertbigdata!(fgl, vert, description, data)

Append big data element into current blob store and update associated global vertex information.

source

appendvertbigdata!(fg, sym, descr, data)

Append big data element into current blob store using parent appendvertbigdata!, but here specified by symbol of variable node in the FactorGraph. Note the default data layer api definition. User must define dlapi to refetching the vertex from the data layer. localapi avoids repeated network database fetches.

source

Caesar.updatenewverts! — Function.

updatenewverts!(fgl::FactorGraph; N::Int)

Convert vertices of session in Neo4j DB with Caesar.jl's required data elements in preparation for MM-iSAMCloudSolve process.

source

Caesar.getVertNeoIDs! — Function.

getVertNeoIDs!(::CloudGraph, res::Dict{Symbol, Int}; session::AbstractString="NA", robot::AbstractString="NA", user::AbstractString="NA")

Insert into and return dict res with Neo4j IDs of ExVertex labels as stored per session in Neo4j database.

source

Caesar.whosNear3D — Function.

whosNear3D(cg, session, robot, user; x, y, z, roll, pitch, yaw, dist, angle)

Find vertices near the point specified and return dictionary of symbol to Neo4j ID pairs.

source

Caesar.getprpt2kde — Function.

getprp2kde(::CloudGraph, neoids::Vector{Int}; N::Int=100)

Return PriorPoint2DensityNH with N points based on beliefs of neoids, and equal share null hypothesis between length(neoids)+1 beliefs.

source

Caesar.whosNear2D — Function.

whosNear2D(cg, session, robot, user; x, y, yaw, dist, angle)

Find vertices near the point specified and return dictionary of symbol to Neo4j ID pairs.

source

Caesar.hasBigDataElement — Function.

hasBigDataElement(vertex, description)

Return true if vertex has bigDataElements with matching description.

source

Caesar.getAllLandmarkNeoIDs — Function.

getAllLandmarkNeoIDs(::Dict{Symbol, Dict{Symbol, Int}}, ::Symbol)

Return Vector{Int} of Neo4j vertex IDs relating to symbol, as listed in lm2others.

source

Caesar.consoleaskuserfordb — Function.

consoleaskuserfordb(; nparticles, drawdepth, clearslamindb, multisession, drawedges)

Obtain database addresses and login credientials from STDIN, as well as a few case dependent options.

source

Caesar.standardcloudgraphsetup — Function.

standardcloudgraphsetup(; addrdict, nparticles, drawdepth, drawedges, clearslamindb, multisession)

Connect to databases via network according to addrdict, or ask user for credentials and return active cloudGraph object, as well as addrdict.

source

Caesar.resetentireremotesession — Function.

resetentireremotesession(conn, session, robot, user)

match (n:session) remove n.backendset, n.ready, n.data, n.bigData, n.label, n.packedType, n.exVertexId, n.shape, n.width set n :NEWDATA return n

source

Caesar.getfirstpose — Function.

getfirstpose(cg::CloudGraph, session::AbstractString, robot::AbstractString, user::AbstractString)

Return Tuple{Symbol, Int} of first pose symbol and Neo4j node ID.

source

Caesar.getExVertexNeoIDs — Function.

getExVertexNeoIDs(conn; label, ready, backendset, session, robot, user, reqbackendset)

Return array of tuples with ExVertex IDs and Neo4j IDs for vertices with label in session.

source

Caesar.findExistingMSConstraints — Function.

findExistingMSConstraints(fgl::FactorGraph)

Return Dict{Symbol, Int} of vertex symbol to Neo4j node ID of MULTISESSION constraints in this fgl.sessionname.

source

Caesar.removeNeo4jID — Function.

removeNeo4jID(cg::CloudGraph, neoid=-1)

Remove node from Neo4j according to Neo4j Node ID. Big data elements that may be associated with this node are not removed.

source

Caesar.getBigDataElement — Function.

getBigDataElement(vertex, description)

Walk through vertex bigDataElements and return the last matching description.

source

Caesar.rmInstMultisessionPriors! — Function.

rmInstMultisessionPriors!(::CloudGraph; session<:AbstractString=, multisessions::Vector{<:AbstractString}= )

source

Caesar.insertrobotdatafirstpose! — Function.

insertrobotdatafirstpose!(cg::CloudGraph, session::AbstractString, robot::AbstractString, user::AbstractString, robotdict::Dict)

Saves robotdict via JSON to first pose in a SESSION in the database. Used for storing general robot specific data in easily accessible manner. Can fetch later retrieve same dict with counterpart fetchrobotdatafirstpose function.

source

Caesar.fetchsubgraph! — Function.

fetchsubgraph!(::FactorGraph, ::Vector{CloudVertex}, numneighbors::Int=0)

Fetch and insert list of CloudVertices into FactorGraph object, up to neighbor depth.

source

fetchsubgraph!(::FactorGraph, ::Vector{Int}, numneighbors::Int=0)

Fetch and insert list of Neo4j IDs into FactorGraph object, up to neighbor depth.

source

Caesar.getLocalSubGraphMultisession — Function.

getLocalSubGraphMultisession{T <: AbstractString}(cg::CloudGraph, lm2others; session::T="", numneighbors::Int=0)

Return subgraph copy of type FactorGraph contaning values from session in lm2others, and Vector{Symbol} of primary key symbols used for graph exstraction.

source

Caesar.getnewvertdict — Function.

getnewvertdict(conn, session::AbstractString, robot::AbstractString, user::AbstractString)

Return a dictionary with frtend and mongo_keys json string information for :NEWDATA elements in Neo4j database.

source

RoME

RoME.getRangeKDEMax2D — Function.

getRangeKDEMax2D(fgl, vsym1, vsym2)

Calculate the cartesian distance between two vertices in the graph using their symbol name, and by maximum belief point.

getRangeKDEMax2D(cgl::CloudGraph, session::AbstractString, vsym1::Symbol, vsym2::Symbol)

Calculate the cartesian distange between two vertices in the graph, by session and symbol names, and by maximum belief point.

source

RoME.initFactorGraph! — Function.

initFactorGraph!(fg; P0, init, N, lbl, ready, firstPoseType, labels)

Initialize a factor graph object as Pose2, Pose3, or neither and returns variable and factor symbols as array.

RoME.addOdoFG! — Function.

addOdoFG!(fg, n, DX, cov; N, ready, labels)

Create a new variable node and insert odometry constraint factor between which will automatically increment latest pose symbol x<k+1> for new node new node and constraint factor are returned as a tuple.

addOdoFG!(fgl, odo; N, ready, labels)

Create a new variable node and insert odometry constraint factor between which will automatically increment latest pose symbol x<k+1> for new node new node and constraint factor are returned as a tuple.

IncrementalInference

IncrementalInference.addNode! — Function.

addNode!(fg, lbl, softtype; N, autoinit, ready, dontmargin, labels, api, uid, smalldata)

Add a node (variable) to a graph. Use this over the other dispatches.

addNode!(fg, lbl, softtype; N, autoinit, ready, dontmargin, labels, api, uid, smalldata)

Add a node (variable) to a graph. Use this over the other dispatches.

IncrementalInference.addFactor! — Function.

addFactor!(fgl, Xi, usrfnc; multihypo, ready, api, labels, uid, autoinit, threadmodel)

Add factor with user defined type <: FunctorInferenceType to the factor graph object. Define whether the automatic initialization of variables should be performed. Use order sensitive multihypo keyword argument to define if any variables are related to data association uncertainty.

addFactor!(fgl, xisyms, usrfnc; multihypo, ready, api, labels, uid, autoinit, threadmodel)

Add factor with user defined type <: FunctorInferenceType to the factor graph object. Define whether the automatic initialization of variables should be performed. Use order sensitive multihypo keyword argument to define if any variables are related to data association uncertainty.

IncrementalInference.allnums — Function.

Test if all elements of the string is a number: Ex, "123" is true, "1_2" is false.

IncrementalInference.approxConv — Function.

approxConv(fgl, fct, towards; api, N)

Draw samples from the approximate convolution of towards symbol using factor fct relative to the other variables. In addition the api can be adjusted to recover the data from elsewhere (likely to be replaced/removed in the future).

IncrementalInference.batchSolve! — Function.

batchSolve!(fgl; drawpdf, show, N, recursive, dbg)

Perform multimodal incremental smoothing and mapping (mm-iSAM) computations over given factor graph fgl::FactorGraph on the local computer. A pdf of the Bayes (Junction) tree will be generated in the working folder with drawpdf=true

IncrementalInference.childCliqs — Function.

childCliqs(treel, cliq)

Return a vector of child cliques to cliq.

IncrementalInference.cliqGibbs — Function.

cliqGibbs(fg, cliq, vertid, inmsgs, N, dbg)

Perform one step of the minibatch clique Gibbs operation for solving the Chapman-Kolmogov trasit integral – here involving separate approximate functional convolution and product operations.

IncrementalInference.convert2packedfunctionnode — Function.

convert2packedfunctionnode(fgl, fsym)
convert2packedfunctionnode(fgl, fsym, api)

Encode complicated function node type to related 'Packed<type>' format assuming a user supplied convert function .

IncrementalInference.decodefg — Function.

decodefg(fgs; api)

Unpack PackedFunctionNodeData formats back to regular FunctonNodeData.

IncrementalInference.doautoinit! — Function.

doautoinit!(fgl, Xi; api, singles, N)

initialize destination variable nodes based on this factor in factor graph, fg, generally called during addFactor!. Destination factor is first (singletons) or second (dim 2 pairwise) variable vertex in Xi.

doautoinit!(fgl, xsyms; api, singles, N)

Initialize destination variable nodes based on this factor in factor graph, fg, generally called during addFactor!. Destination factor is first (singletons) or second (dim 2 pairwise) variable vertex in Xi.

IncrementalInference.downMsgPassingRecursive — Function.

downMsgPassingRecursive(inp; N, dbg, drawpdf)

Pass NBPMessages back down the tree – pre order tree traversal.

IncrementalInference.dwnMsg — Function.

dwnMsg(cliq)

Return the last down message stored in cliq of Bayes (Junction) tree.

IncrementalInference.emptyFactorGraph — Function.

emptyFactorGraph(; reference)

Construct an empty FactorGraph object with the minimum amount of information / memory populated.

IncrementalInference.encodefg — Function.

encodefg(fgl; api)

Make a full memory copy of the graph and encode all composite function node types – assuming that convert methods for 'Packed<type>' formats exist. The same converters are used for database persistence with CloudGraphs.jl.

IncrementalInference.fifoFreeze! — Function.

fifoFreeze!(fgl)

Freeze nodes that are older than the quasi fixed-lag length defined by fg.qfl, according to fg.fifo ordering.

Future:

Allow different freezing strategies beyond fifo.

IncrementalInference.findRelatedFromPotential — Function.

findRelatedFromPotential(fg, idfct, vertid, N)
findRelatedFromPotential(fg, idfct, vertid, N, dbg)

Compute proposal belief on vertid through idfct representing some constraint in factor graph. Always full dimension variable node – partial constraints will only influence subset of variable dimensions. The remaining dimensions will keep pre-existing variable values.

IncrementalInference.fmcmc! — Function.

fmcmc!(fgl, cliq, fmsgs, IDs, N, MCMCIter)
fmcmc!(fgl, cliq, fmsgs, IDs, N, MCMCIter, dbg)

Iterate successive approximations of clique marginal beliefs by means of the stipulated proposal convolutions and products of the functional objects for tree clique cliq.

IncrementalInference.getCurrentWorkspaceFactors — Function.

getCurrentWorkspaceFactors()

Return all factors currently registered in the workspace.

IncrementalInference.getCurrentWorkspaceVariables — Function.

getCurrentWorkspaceVariables()

Return all variables currently registered in the workspace.

IncrementalInference.getKDE — Function.

getKDE(v)

Get KernelDensityEstimate kde estimate stored in variable node.

IncrementalInference.getVertKDE — Function.

getVertKDE(v)

Get KernelDensityEstimate kde estimate stored in variable node.

IncrementalInference.initializeNode! — Function.

initializeNode!(fgl, sym; N, api)

Initialize the belief of a variable node in the factor graph struct.

IncrementalInference.landmarks — Function.

landmarks(fgl::FactorGraph, vsym::Symbol)

Return Vector{Symbol} of landmarks attached to vertex vsym in fgl.

IncrementalInference.loadjld — Function.

loadjld(; file)

Opposite of savejld(fg, gt=gt, file="tempfg.jl") to load data from file. This function uses the unpacking converters for converting all PackedInferenceType to FunctorInferenceType.

IncrementalInference.localProduct — Function.

localProduct(fgl, sym; N, dbg, api)

Using factor graph object fg, project belief through connected factors (convolution with conditional) to variable sym followed by a approximate functional product.

Return: product belief, full proposals, partial dimension proposals, labels

IncrementalInference.ls — Function.

ls(fgl, lbl; api, ring)

ls(fgl, lbls; api, ring)

Experimental union of elements version of ls(::FactorGraph, ::Symbol). Not mean't to replace broadcasting ls.(fg, [:x1;:x2])

ls(fgl; key1, key2)

List the nodes in a factor graph.

Examples

ls(fg)

ls(cgl, session, robot, user; sym, neoid, exvid)

List neighbors to node in cgl::CloudGraph by returning Dict{Sym}=(exvid, neoid, Symbol[labels]), and can take any of the three as input node identifier. Not specifying an identifier will result in all Variable nodes being returned.

source

IncrementalInference.lsf — Function.

lsf(fgl, lbl; api)

List factors in a factor graph.

Examples

lsf(fg)

IncrementalInference.lsRear — Function.

lsRear(fgl)
lsRear(fgl, n)

Return array of all variable nodes connected to the last n many poses (:x*).

Example:

# Shallow copy the tail end of poses from a factor graph `fg1`
vars = lsRear(fg1, 5)
fg1_r5 = subgraphFromVerts(fg1, vars)

IncrementalInference.manualinit! — Function.

manualinit!(fgl, sym, usefcts)

Workaround function when first-version (factor graph based) auto initialization fails. Usually occurs when using factors that have high connectivity to multiple variables.

IncrementalInference.packFromLocalPotentials! — Function.

packFromLocalPotentials!(fgl, dens, wfac, cliq, vertid, N)
packFromLocalPotentials!(fgl, dens, wfac, cliq, vertid, N, dbg)

Add all potentials associated with this clique and vertid to dens.

IncrementalInference.printgraphmax — Function.

printgraphmax(fgl)

Print the maximum point values form all variables approximate marginals in the factor graph. The full marginal can be recovered for example X0 = getVertKDE(fg, :x0).

IncrementalInference.productpartials! — Function.

productpartials!(pGM, dummy, partials)

Multiply different dimensions from partial constraints individually.

IncrementalInference.prodmultiplefullpartials — Function.

prodmultiplefullpartials(dens, partials, Ndims, N)

Multiply various full and partial dimension constraints.

IncrementalInference.prodmultipleonefullpartials — Function.

prodmultipleonefullpartials(dens, partials, Ndims, N)

Multiply a single full and several partial dimension constraints.

IncrementalInference.savejld — Function.

savejld(fgl; file, groundtruth)

Save mostly complete Factor Graph type by converting complicated FunctionNodeData types to 'Packed' types using user supplied converters. Ground truth can also be saved and recovered by the associated loadjld(file="tempfg.jld2") method.

IncrementalInference.setDwnMsg! — Function.

setDwnMsg!(cliql, msgs)

Set the downward passing message for Bayes (Junction) tree clique cliql.

IncrementalInference.setfreeze! — Function.

setfreeze!(fgl, sym)

Set variable(s) sym of factor graph to be marginalized – i.e. not be updated by inference computation.

IncrementalInference.setUpMsg! — Function.

setUpMsg!(cliql, msgs)

Set the upward passing message for Bayes (Junction) tree clique cliql.

IncrementalInference.subgraphFromVerts — Function.

subgraphFromVerts(fgl, verts; neighbors)

Explore all shortest paths combinations in verts, add neighbors and reference subgraph using unique index into graph data structure.

IncrementalInference.treeProductDwn — Function.

treeProductDwn(fg, tree, cliq, sym; N, dbg)

Calculate a fresh–-single step–-approximation to the variable sym in clique cliq as though during the downward message passing. The full inference algorithm may repeatedly calculate successive apprimxations to the variable based on the structure of variables, factors, and incoming messages to this clique. Which clique to be used is defined by frontal variable symbols (cliq in this case) – see whichCliq(...) for more details. The sym symbol indicates which symbol of this clique to be calculated. Note that the sym variable must appear in the clique where cliq is a frontal variable.

IncrementalInference.treeProductUp — Function.

treeProductUp(fg, tree, cliq, sym; N, dbg)

Calculate a fresh–-single step–-approximation to the variable sym in clique cliq as though during the upward message passing. The full inference algorithm may repeatedly calculate successive apprimxations to the variable based on the structure of variables, factors, and incoming messages to this clique. Which clique to be used is defined by frontal variable symbols (cliq in this case) – see whichCliq(...) for more details. The sym symbol indicates which symbol of this clique to be calculated. Note that the sym variable must appear in the clique where cliq is a frontal variable.

IncrementalInference.updateFGBT! — Function.

updateFGBT!(fg, bt, cliqID, ddt; dbg, fillcolor)

Update cliq cliqID in Bayes (Juction) tree bt according to contents of ddt – intended use is to update main clique after a downward belief propagation computation has been completed per clique.

updateFGBT!(fg, bt, cliqID, urt; dbg, fillcolor)

Update cliq cliqID in Bayes (Juction) tree bt according to contents of urt – intended use is to update main clique after a upward belief propagation computation has been completed per clique.

IncrementalInference.upGibbsCliqueDensity — Function.

upGibbsCliqueDensity(inp)
upGibbsCliqueDensity(inp, N)
upGibbsCliqueDensity(inp, N, dbg)

Perform computations required for the upward message passing during belief propation on the Bayes (Junction) tree. This function is usually called as part via remote_call for multiprocess dispatch.

IncrementalInference.upMsg — Function.

upMsg(cliq)

Return the last up message stored in cliq of Bayes (Junction) tree.

IncrementalInference.writeGraphPdf — Function.

writeGraphPdf(fgl; viewerapp, filepath, engine, show)

Export a dot and pdf file drawn by Graphviz showing the factor graph.