Visualization Function Reference

Visualization Function Reference

RoMEPlotting

RoMEPlotting

RoMEPlotting.drawLandms — Function

drawLandms(fg; from, to, minnei, meanmax, lbls, showmm, drawhist, c, MM)

2D plot of landmarks, assuming :l1, :l2, ... :ln. Use from and to to control the range of landmarks n to include.

RoMEPlotting.drawPoses — Function

2D plot of all poses, assuming poses are labeled from `::Symbol type :x0, :x1, ..., :xn. Use to and from to limit the range of numbers n to be drawn. The underlying histogram can be enabled or disabled, and the size of maximum-point belief estimate cursors can be controlled with spscale.

Future:

Relax to user defined pose labeling scheme, for example :p1, :p2, ...

RoMEPlotting.drawPosesLandms — Function

2D plot of both poses and landmarks contained in factor graph. Assuming poses and landmarks are labeled :x1, :x2, ... and :l0, :l1, ..., respectively. The rnage of numbers to include can be controlled with from and to along with other keyword functionality for manipulating the plot.

Notes

assumes :l1, :l2, ... for landmarks – not using tags=[:LANDMARK] here yet (TODO).

KernelDensityEstimatePlotting.plotKDE — Function

A peneric KDE plotting function that allows marginals of higher dimensional beliefs and various keyword options.

Example:

p = kde!(randn(3,100))

plotKDE(p)
plotKDE(p, dims=[1;2], levels=3)
plotKDE(p, dims=[1])

q = kde!(5*randn(3,100))
plotKDE([p;q])
plotKDE([p;q], dims=[1;2], levels=3)
plotKDE([p;q], dims=[1])

RoMEPlotting.plotPose — Function

plotPose(pt, pp)
plotPose(pt, pp, title; levels, c, axis, scale)

Plot pose belief as contour information on visually sensible manifolds.

plotPose(fgl, syms; levels, c, axis, show, filepath, app)

Example: pl = plotPose(fg, [:x1; :x2; :x3])

RoMEPlotting.plotProductVsKDE — Function

Analysis function to compare KDE plots between the factor graph centric product of a variable with current value stored in the factor graph object.

WORK IN PROGRESS Not all functions have been added to this directory yet.