Visualization Reference

Visualization Function Reference

Visualization Function Reference

RoMEPlotting

KernelDensityEstimatePlotting

RoMEPlotting

RoMEPlotting.drawLandms — Function.

drawLandms(fg; from, to, minnei, meanmax, lbls, showmm, drawhist, c, MM, api)

2D plot of landmarks, assuming :l1, :l2, ... :ln. Use from and to to control the range of landmarks n to include.

RoMEPlotting.drawPoses — Function.

drawPoses(fg; from, to, meanmax, lbls, drawhist, spscale, api)

2D plot of all poses, assuming poses are labeled from `::Symbol type :x0, :x1, ..., :xn. Use to and from to limit the range of numbers n to be drawn. The underlying histogram can be enabled or disabled, and the size of maximum-point belief estimate cursors can be controlled with spscale.

Future:

Relax to user defined pose labeling scheme, for example :p1, :p2, ...

RoMEPlotting.drawPosesLandms — Function.

drawPosesLandms(fgl; from, to, minnei, meanmax, lbls, drawhist, MM, showmm, spscale, window, api, xmin, xmax, ymin, ymax)

2D plot of both poses and landmarks contained in factor graph. Assuming poses and landmarks are labeled :x1, :x2, ... and :l0, :l1, ..., respectively. The rnage of numbers to include can be controlled with from and to along with other keyword functionality for manipulating the plot.

KernelDensityEstimatePlotting.plotKDE — Function.

plotKDE(fgl, sym; axis, dims, c, levels, title)

A peneric KDE plotting function that allows marginals of higher dimensional beliefs and various keyword options.

Example:

p = kde!(randn(3,100))

plotKDE(p)
plotKDE(p, dims=[1;2], levels=3)
plotKDE(p, dims=[1])

q = kde!(5*randn(3,100))
plotKDE([p;q])
plotKDE([p;q], dims=[1;2], levels=3)
plotKDE([p;q], dims=[1])

RoMEPlotting.plotProductVsKDE — Function.

plotProductVsKDE(fgl, sym; levels, c)

Analysis function to compare KDE plots between the factor graph centric product of a variable with current value stored in the factor graph object.

KernelDensityEstimatePlotting

plotKDE

WORK IN PROGRESS Not all functions have been added to this directory yet.